fix demo因权限问题导致下载失败

+++ app/src/main/java/com/qw/example/widget/QProgress.java
@@ -22,8 +22,8 @@ public class QProgress extends View {
     private Paint mPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
     private RectF mRectF;
 
-    private int max = 100;
-    private int progress = 0;
+    private long max = 100;
+    private long progress = 0;
 
     private int startX;
     private int endX;
